Server should now connect to websocket
even if its behind a virtual host. Sometimes. Situationally. It gets the
IP of the server, not the host name. If the "real" server is reachable
from the internet, this will work. If not, figure out how to forward
websocket requests on your own. (apache won't do it happily).

Also, removed the word "fuck", although that may have been the previous
commit.